added update gender and height in mainService loadState

+++ app/src/main/java/se/kth/projectarbor/project_arbor/Pedometer.java
@@ -146,7 +146,7 @@ class Pedometer {
     }
 
     void setHeight(float height){
-        this.height = height;
+        this.height = (double) height;
     }
 
     public void reset() {
